fixed errors for V1
